Moved classes to different folders.
