+++ Assets/Scripts/NewGame/Components.meta
@@ -1,7 +1,7 @@
 fileFormatVersion: 2
-guid: 62b396f00c48fea4d9d6eb616f2ca328
+guid: 8470c05a608445a41ab53a39b2b9ffcc
 folderAsset: yes
-timeCreated: 1467472744
+timeCreated: 1467544486
 licenseType: Free
 DefaultImporter:
   userData:
