+++ Assets/Scripts/NewGame/Components/NormalBlockComponent.cs
