+++ Assets/Scripts/NewGame/Components/PaddleComponent.cs
