+++ Assets/Scripts/NewGame/Entities.meta
@@ -1,7 +1,7 @@
 fileFormatVersion: 2
-guid: ecf168b82a2799f4da4b8490384625af
+guid: 4ca5fa4523f5bc2439d353c431f08667
 folderAsset: yes
-timeCreated: 1467472446
+timeCreated: 1467544424
 licenseType: Free
 DefaultImporter:
   userData:
