--- Assets/Scripts/NewGame/Entities/Ball.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 5e5279445daa19147a90224b38d7c630
-folderAsset: yes
-timeCreated: 1467472561
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
