--- Assets/Scripts/NewGame/Entities/DeadZone.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 0108ff6814d273749a2f9648c6a1d65e
-folderAsset: yes
-timeCreated: 1467489990
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
