--- Assets/Scripts/NewGame/Entities/DeadZone/DeadZone.cs
@@ -1,8 +1,0 @@
-﻿
-namespace Game
-{
-    public interface DeadZone {
-        void CollidedWithBall(Ball ball);
-    }
-}
-
