+++ Assets/Scripts/NewGame/Entities/DefaultPaddle.cs
