+++ Assets/Scripts/NewGame/Interfaces.meta
@@ -1,7 +1,7 @@
 fileFormatVersion: 2
-guid: a99dc56ff14a6384580a9805c784ff47
+guid: 75a03d7d77a99664b8a6da4810972347
 folderAsset: yes
-timeCreated: 1467472860
+timeCreated: 1467544386
 licenseType: Free
 DefaultImporter:
   userData:
