+++ Assets/Scripts/NewGame/Interfaces/GameEntity.cs.meta
