+++ Assets/Scripts/NewGame/Interfaces/Paddle.cs
